needed explicit integer division in python 3

+++ qq_code/mcmc_2disks/data_prepare/prepare_data_jackknife.py
@@ -39,7 +39,7 @@ def main():
         # make jackknife samples
         N_jackknife = config.N_jackknife # number of jackknife samples
 
-        N = len(star_list) / N_jackknife
+        N = int(len(star_list) / N_jackknife)
 
         for i in range(N_jackknife):
             output_list = star_list[:(N * i)] + star_list[(N * i + N):]
